+++ lib/githubGraph.py
@@ -4,7 +4,8 @@
 import time
 import datetime
 import requests
-import utils
+from utils import printf, getPassword, getDatetime
+from metprint import LogType
 
 def getGithubApiRequest(query, variables=None, jsonOnly=True):
 	"""use this to get json from api (returns some data to module variables)
@@ -17,15 +18,15 @@ def getGithubApiRequest(query, variables=None, jsonOnly=True):
 		"query": query},
 		headers={
 			"Authorization": "bearer " +
-			 utils.getPassword()})
+			 getPassword()})
 	if int(request.headers["X-RateLimit-Remaining"]) < 1:
-		utils.logPrint("Remaining rate limit is zero. Try again at {}"
+		printf.logPrint("Remaining rate limit is zero. Try again at {}"
 		.format(str(time.ctime(request.headers["X-RateLimit-Reset"]))), "error")
 
 	requestJson = request.json()
 	if "message" in requestJson:
-		utils.logPrint("Some error has occurred", "error")
-		utils.logPrint(requestJson)
+		printf.logPrint("Some error has occurred", "error")
+		printf.logPrint(requestJson)
 
 	return requestJson if jsonOnly else request
 
@@ -73,7 +74,7 @@ def getListOfAliveForks(repoData, lifespan, enableNewer=True):
 	aliveRepos = []
 	for forkedRepo in forkedRepos:
 		isAlive = sourceAlive(forkedRepo, lifespan)
-		isNewer = utils.getDatetime(forkedRepo["pushedAt"]) > utils.getDatetime(repoData["pushedAt"])
+		isNewer = getDatetime(forkedRepo["pushedAt"]) > getDatetime(repoData["pushedAt"])
 		if (isAlive and isNewer) or (isAlive and not enableNewer):
 			aliveRepos.append(forkedRepo)
 	return aliveRepos, forkedRepos
@@ -144,22 +145,22 @@ def getUserGists(username):
 		{"login": username})["data"]["user"]["gists"]["nodes"]
 
 
-
-def getListOfUserRepos(username, context="repositories", lifespan=520):
+def getListOfRepos(login, context="repositories", organization=False, lifespan=520):
 	"""Get a list of repos using a username and type: "repositories" (user public repos),
 	"watching" (user watching), "starredRepositories" (stars)
 	"""
 	repos = []
 	hasNextPage = True
 	after = ""
+	starredOrPushed = "STARRED_AT" if context == "starredRepositories" else "PUSHED_AT"
+	userOrOrg = "organization" if organization else "user"
 	while hasNextPage:
 		includeIfAfter = """after:"$after",""" if after != "" else ""
-		starredAtIfStarred = "STARRED_AT" if context == "starredRepositories" else "PUSHED_AT"
 		repoPage = getGithubApiRequest("""
 		query {
-			user(login:"$login") {
+			""" + userOrOrg + """(login:"$login") {
 				$context(first:100, """ + includeIfAfter +
-				"""orderBy:{direction:DESC, field:""" + starredAtIfStarred + """}){
+				"""orderBy:{direction:DESC, field:""" + starredOrPushed + """}){
 					pageInfo {
 						hasNextPage
 						endCursor
@@ -177,7 +178,7 @@ def getListOfUserRepos(username, context="repositories", lifespan=520):
 					}
 				}
 			}""",
-		{"login": username, "context": context, "after": after})["data"]["user"][context]
+		{"login": login, "context": context, "after": after})["data"][userOrOrg][context]
 		repos.extend(repoPage["nodes"])
 		hasNextPage = repoPage["pageInfo"]["hasNextPage"] and sourceAlive(repoPage["nodes"][99], lifespan)
 		after = repoPage["pageInfo"]["endCursor"]
@@ -186,39 +187,39 @@ def getListOfUserRepos(username, context="repositories", lifespan=520):
 
 def printIssue(issue):
 	'''Print issue function '''
-	utils.logPrint(("[\033[91mClosed\033[00m] " if issue["state"] == "closed" else "")
-	+ issue["title"], "bold")
-	utils.logPrint(issue["pushedAt"])
+	printf.logPrint(("[\033[91mClosed\033[00m] " if issue["state"] == "closed" else "")
+	+ issue["title"], LogType.BOLD)
+	printf.logPrint(issue["pushedAt"])
 
 def printUser(user):
 	'''Print user function '''
-	utils.logPrint(user["login"], "bold")
-	utils.logPrint(user["url"])
+	printf.logPrint(user["login"], LogType.BOLD)
+	printf.logPrint(user["url"])
 
 def printGist(gist):
 	'''Print gist function '''
-	utils.logPrint(gist["description"], "bold")
-	utils.logPrint("Files: {}" .format([gFile['name'] for gFile in gist["files"]]), "bold")
-	utils.logPrint(gist["url"])
+	printf.logPrint(gist["description"], LogType.BOLD)
+	printf.logPrint("Files: {}" .format([gFile['name'] for gFile in gist["files"]]), LogType.BOLD)
+	printf.logPrint(gist["url"])
 
 def printRepo(repo):
 	'''Print repo function '''
 	if all(key in repo for key in ("isArchived", "name")):
-		utils.logPrint("{}"
-		.format(("[\033[91mArchived\033[00m] " if repo["isArchived"] else "") + repo["name"]), "bold")
+		printf.logPrint("{}"
+		.format(("[\033[91mArchived\033[00m] " if repo["isArchived"] else "") + repo["name"]), LogType.BOLD)
 	else:
 		return
 	description = repo["description"] if "description" in repo else "[description]"
 	language = repo["primaryLanguage"]["name"] if repo["primaryLanguage"] is not None else "[unknown]"
 	licenseName = repo["licenseInfo"]["name"] if repo["licenseInfo"] is not None else "[unknown]"
 	pushed = repo["pushedAt"] if "pushedAt" in repo else "[unknown]"
-	utils.logPrint("{}\nLanguage: {}, License: {}, Last Pushed: {}"
+	printf.logPrint("{}\nLanguage: {}, License: {}, Last Pushed: {}"
 	.format(description, language, licenseName, pushed))
-	utils.logPrint("Link: {}" .format(repo["url"]))
+	printf.logPrint("Link: {}" .format(repo["url"]))
 
 
 def sourceAlive(repoData, lifespan):
 	"""Is source repo alive?
 	"""
-	return utils.getDatetime(repoData["pushedAt"]) > (datetime.datetime.now() -
+	return getDatetime(repoData["pushedAt"]) > (datetime.datetime.now() -
 	datetime.timedelta(weeks=lifespan))