+++ lib/utils.py
@@ -3,6 +3,9 @@
 import os
 import datetime
 import json
+from metprint import LogType, Logger, FHFormatter
+
+printf = Logger(FHFormatter())
 
 
 def clear():
@@ -14,15 +17,6 @@ def getDatetime(datetimeIn):
 	"""Get the datetime from a date in the format YYYY-MM-DDThh:mm:ssZ e.g. 2000-01-01T00:00:00Z
 	"""
 	return datetime.datetime.strptime(datetimeIn, "%Y-%m-%dT%H:%M:%SZ")
-
-
-def logPrint(printText, printType="standard"):
-	"""use this to print. standard, success, warning, error, info, bold
-	"""
-	types = {"standard": "{}", "success": "[\033[92m+ Success\033[00m] {}",
-	"warning": "[\033[93m/ Warning\033[00m] {}", "error": "[\033[91m- Error\033[00m] {}",
-	"info": "[\033[96m* Info\033[00m] {}", "bold": "\033[01m{}\033[00m"}
-	print(types[printType.lower()] .format(printText))
 
 
 def getUsername():
@@ -41,7 +35,7 @@ def getUsernameAndLifespan():
 	try:
 		lifespan = int(input("Set the repo lifespan (weeks - eg. 1 - default=36)\n>"))
 	except ValueError:
-		logPrint("Invalid input - using default", "warning")
+		printf.logPrint("Invalid input - using default", LogType.WARNING)
 	return username, lifespan
 
 
@@ -49,11 +43,11 @@ if 'AUTH' not in locals():
 	try:
 		authJson = json.loads(open("env.json", "r").read())
 		AUTH = (authJson["username"], authJson["password"])
-		logPrint("Authenticated user {}" .format(authJson["username"]), "success")
+		printf.logPrint("Authenticated user {}" .format(authJson["username"]), LogType.SUCCESS)
 	except FileNotFoundError:
-		logPrint(
+		printf.logPrint(
 		"Not authenticated - Do you want to log in? (just hit enter if not)",
-		"warning")
+		LogType.WARNING)
 		AUTH = (input("Enter your username\n>"), input("Enter your password\n>"))
 		if len(AUTH[0]) == 0 or len(AUTH[1]) == 0:
-			logPrint("Not authenticated - rate limit is 60 requests per hour", "warning")
+			printf.logPrint("Not authenticated - rate limit is 60 requests per hour", LogType.WARNING)