+++ main/UserReposActive.py
@@ -8,49 +8,50 @@ from pathlib import Path
 THISDIR = str(Path(__file__).resolve().parent)
 sys.path.insert(0, os.path.dirname(THISDIR) + "/lib")
 
+from metprint import LogType
 import githubGraph
-import utils
+from utils import printf, getUsernameAndLifespan
 
 def forEachRepo(sourceRepo):
 	"""Is source repo alive?
 	"""
-	printStr = ["dead", "error"]
+	printStr = ["dead", LogType.ERROR]
 	if githubGraph.sourceAlive(sourceRepo, death):
-		printStr = ["alive", "success"]
-	utils.logPrint("Source repo is {}! Head to {}"
+		printStr = ["alive", LogType.SUCCESS]
+	printf.logPrint("Source repo is {}! Head to {}"
 	.format(printStr[0], sourceRepo["url"]), printStr[1])
 
 	"""Get list of forked repos that are alive and newer than the source repo
 	"""
 	aliveRepos, forkedRepos = githubGraph.getListOfAliveForks(repo, death)
 
-	utils.logPrint("{} out of {} Forked repos are alive and newer than the source!"
-	.format(len(aliveRepos), len(forkedRepos)), "bold")
+	printf.logPrint("{} out of {} Forked repos are alive and newer than the source!"
+	.format(len(aliveRepos), len(forkedRepos)), LogType.BOLD)
 	for aliveRepo in aliveRepos:
 		githubGraph.printRepo(aliveRepo)
 
 
-username, death = utils.getUsernameAndLifespan()
+username, death = getUsernameAndLifespan()
 
 
 choice = input("User repos, watched or starred (R/w/s)>")
 if choice.lower() == "s":
 	"""Get list of user starred
 	"""
-	starredRepos = githubGraph.getListOfUserRepos(username, "starredRepositories")
+	starredRepos = githubGraph.getListOfRepos(username, "starredRepositories")
 	for repo in starredRepos:
 		forEachRepo(repo)
 
 elif choice.lower() == "w":
 	"""Get list of user watched
 	"""
-	watchedRepos = githubGraph.getListOfUserRepos(username, "watching")
+	watchedRepos = githubGraph.getListOfRepos(username, "watching")
 	for repo in watchedRepos:
 		forEachRepo(repo)
 
 else:
 	"""Get list of user repos
 	"""
-	sourceRepos = githubGraph.getListOfUserRepos(username, "repositories")
+	sourceRepos = githubGraph.getListOfRepos(username, "repositories")
 	for repo in sourceRepos:
 		forEachRepo(repo)